update docs script
